# grotto/backend/openapi.yaml
openapi: 3.1.0
info:
  title: Grotto
  description: A general application for joining the features of Sonarr and Radarr
    into a single unified platform.
  version: 1.0.0
paths:
  /api/auth/register:
    post:
      tags:
      - auth
      summary: Register
      description: "Register a new user account.\n\nArgs:\n    user_data: UserCreate\
        \ schema with email, username, and password.\n    db: Database session.\n\n\
        Returns:\n    UserResponse with the created user's details.\n\nRaises:\n \
        \   HTTPException: 400 Bad Request if email or username already exists."
      operationId: register_api_auth_register_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreate'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/auth/login:
    post:
      tags:
      - auth
      summary: Login
      description: "Authenticate user with email and password, return JWT access token.\n\
        \nArgs:\n    login_data: LoginRequest schema with email and password.\n  \
        \  db: Database session.\n\nReturns:\n    Token with access_token and token_type\
        \ (\"bearer\").\n\nRaises:\n    HTTPException: 401 Unauthorized if email/password\
        \ combination is invalid."
      operationId: login_api_auth_login_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Token'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/auth/me:
    get:
      tags:
      - auth
      summary: Read Users Me
      description: "Get the current authenticated user's profile information.\n\n\
        Requires valid JWT token in Authorization header.\n\nArgs:\n    current_user:\
        \ Current authenticated user (injected via get_current_user).\n\nReturns:\n\
        \    UserResponse with the authenticated user's details.\n\nRaises:\n    HTTPException:\
        \ 401 Unauthorized if no valid token provided."
      operationId: read_users_me_api_auth_me_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
      security:
      - OAuth2PasswordBearer: []
  /api/auth/request-password-reset:
    post:
      tags:
      - auth
      summary: Request Password Reset
      description: "Request a password reset token for the given email address.\n\n\
        If an account exists, a reset token is generated (caller should send reset\
        \ link via email).\nReturns success message regardless to avoid email enumeration.\n\
        \nArgs:\n    email: Email address associated with the account.\n    db: Database\
        \ session.\n\nReturns:\n    JSON message indicating reset link will be sent\
        \ if account exists.\n\nNote:\n    Always returns success message to prevent\
        \ user enumeration attacks."
      operationId: request_password_reset_api_auth_request_password_reset_post
      parameters:
      - name: email
        in: query
        required: true
        schema:
          type: string
          format: email
          title: Email
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/auth/reset-password:
    post:
      tags:
      - auth
      summary: Reset Password
      description: "Reset a user's password using a valid reset token.\n\nArgs:\n\
        \    token: Password reset token (obtained from password reset email).\n \
        \   new_password: New password to set for the account.\n    db: Database session.\n\
        \nReturns:\n    JSON message confirming password has been reset.\n\nRaises:\n\
        \    HTTPException: 400 Bad Request if token is invalid, expired, or not found."
      operationId: reset_password_api_auth_reset_password_post
      parameters:
      - name: token
        in: query
        required: true
        schema:
          type: string
          title: Token
      - name: new_password
        in: query
        required: true
        schema:
          type: string
          title: New Password
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/health:
    get:
      tags:
      - Health
      summary: Health Check
      description: "Health check endpoint for API availability monitoring.\n\nCan\
        \ be used for:\n- Load balancer health checks\n- Kubernetes/container orchestration\
        \ liveness probes\n- Monitoring and alerting systems\n- Uptime verification\n\
        \nReturns:\n    JSON object with status \"healthy\" if API is running."
      operationId: health_check_api_health_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
  /api/media/search:
    get:
      tags:
      - media
      summary: Search Media
      description: "Search Sonarr and Radarr for a title.\n\nArgs:\n    term: Title\
        \ (or partial title) to search for.\n\nReturns:\n    Combined TV and movie\
        \ matches.\n\nRaises:\n    HTTPException: 502 if Sonarr/Radarr can't be reached."
      operationId: search_media_api_media_search_get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: term
        in: query
        required: true
        schema:
          type: string
          title: Term
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MediaSearchResultResponse'
                title: Response Search Media Api Media Search Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/media/library:
    get:
      tags:
      - media
      summary: List Library
      description: "Return everything currently in the Sonarr and Radarr libraries.\n\
        \nRaises:\n    HTTPException: 502 if Sonarr/Radarr can't be reached."
      operationId: list_library_api_media_library_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/MediaResultResponse'
                type: array
                title: Response List Library Api Media Library Get
      security:
      - OAuth2PasswordBearer: []
  /api/media/library/{media_type}/{id}:
    get:
      tags:
      - media
      summary: Get Library Item
      description: "Return the details of a single library item, including its current\n\
        download status if Sonarr/Radarr has it queued or downloading.\n\nArgs:\n\
        \    media_type: Whether `id` refers to a series (Sonarr) or movie (Radarr).\n\
        \    id: The Sonarr or Radarr internal id, as returned by /media/library.\n\
        \nRaises:\n    HTTPException: 502 if Sonarr/Radarr can't be reached or the\
        \ item doesn't exist."
      operationId: get_library_item_api_media_library__media_type___id__get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: media_type
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/MediaType'
      - name: id
        in: path
        required: true
        schema:
          type: integer
          title: Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MediaResultResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/media/add:
    post:
      tags:
      - media
      summary: Add Media
      description: "Add a previously-searched title to the correct library and start\n\
        downloading it.\n\nSearch results aren't cached server-side, so this re-runs\
        \ the search\nfor `term` to locate the exact match by `media_type`/`external_id`\n\
        before adding it.\n\nRaises:\n    HTTPException: 404 if the match can no longer\
        \ be found, 502 if the\n        backend rejects the add request."
      operationId: add_media_api_media_add_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddMediaRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MediaResultResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - OAuth2PasswordBearer: []
  /api/media/remove:
    post:
      tags:
      - media
      summary: Remove Media
      description: "Remove a library item from whichever backend it belongs to.\n\n\
        Raises:\n    HTTPException: 404 if no matching library item is found, 502\
        \ if the\n        backend rejects the remove request."
      operationId: remove_media_api_media_remove_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RemoveMediaRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - OAuth2PasswordBearer: []
components:
  schemas:
    AddMediaRequest:
      properties:
        media_type:
          $ref: '#/components/schemas/MediaType'
        term:
          type: string
          title: Term
        external_id:
          type: integer
          title: External Id
        quality_profile_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Quality Profile Id
        root_folder_path:
          anyOf:
          - type: string
          - type: 'null'
          title: Root Folder Path
        monitored:
          type: boolean
          title: Monitored
          default: true
        search_now:
          type: boolean
          title: Search Now
          default: true
      type: object
      required:
      - media_type
      - term
      - external_id
      title: AddMediaRequest
    DownloadStatusResponse:
      properties:
        status:
          type: string
          title: Status
        tracked_download_state:
          anyOf:
          - type: string
          - type: 'null'
          title: Tracked Download State
        tracked_download_status:
          anyOf:
          - type: string
          - type: 'null'
          title: Tracked Download Status
        title:
          anyOf:
          - type: string
          - type: 'null'
          title: Title
        size:
          type: number
          title: Size
        size_remaining:
          anyOf:
          - type: number
          - type: 'null'
          title: Size Remaining
        time_remaining:
          anyOf:
          - type: string
          - type: 'null'
          title: Time Remaining
        estimated_completion:
          anyOf:
          - type: string
          - type: 'null'
          title: Estimated Completion
        protocol:
          anyOf:
          - type: string
          - type: 'null'
          title: Protocol
        download_client:
          anyOf:
          - type: string
          - type: 'null'
          title: Download Client
        indexer:
          anyOf:
          - type: string
          - type: 'null'
          title: Indexer
        error_message:
          anyOf:
          - type: string
          - type: 'null'
          title: Error Message
      type: object
      required:
      - status
      - tracked_download_state
      - tracked_download_status
      - title
      - size
      - size_remaining
      - time_remaining
      - estimated_completion
      - protocol
      - download_client
      - indexer
      - error_message
      title: DownloadStatusResponse
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    LoginRequest:
      properties:
        email:
          type: string
          format: email
          title: Email
        password:
          type: string
          title: Password
      type: object
      required:
      - email
      - password
      title: LoginRequest
    MediaResultResponse:
      properties:
        media_type:
          $ref: '#/components/schemas/MediaType'
        id:
          type: integer
          title: Id
        title:
          type: string
          title: Title
        monitored:
          type: boolean
          title: Monitored
        external_id:
          type: integer
          title: External Id
        year:
          anyOf:
          - type: integer
          - type: 'null'
          title: Year
        overview:
          anyOf:
          - type: string
          - type: 'null'
          title: Overview
        poster_url:
          anyOf:
          - type: string
          - type: 'null'
          title: Poster Url
        fanart_url:
          anyOf:
          - type: string
          - type: 'null'
          title: Fanart Url
        genres:
          items:
            type: string
          type: array
          title: Genres
        runtime:
          anyOf:
          - type: integer
          - type: 'null'
          title: Runtime
        rating:
          anyOf:
          - type: number
          - type: 'null'
          title: Rating
        network:
          anyOf:
          - type: string
          - type: 'null'
          title: Network
        studio:
          anyOf:
          - type: string
          - type: 'null'
          title: Studio
        download_status:
          anyOf:
          - $ref: '#/components/schemas/DownloadStatusResponse'
          - type: 'null'
      type: object
      required:
      - media_type
      - id
      - title
      - monitored
      - external_id
      - year
      - overview
      - poster_url
      - fanart_url
      - genres
      - runtime
      - rating
      - network
      - studio
      - download_status
      title: MediaResultResponse
    MediaSearchResultResponse:
      properties:
        media_type:
          $ref: '#/components/schemas/MediaType'
        title:
          type: string
          title: Title
        year:
          anyOf:
          - type: integer
          - type: 'null'
          title: Year
        overview:
          anyOf:
          - type: string
          - type: 'null'
          title: Overview
        already_in_library:
          type: boolean
          title: Already In Library
        external_id:
          type: integer
          title: External Id
        poster_url:
          anyOf:
          - type: string
          - type: 'null'
          title: Poster Url
        fanart_url:
          anyOf:
          - type: string
          - type: 'null'
          title: Fanart Url
        genres:
          items:
            type: string
          type: array
          title: Genres
        runtime:
          anyOf:
          - type: integer
          - type: 'null'
          title: Runtime
        rating:
          anyOf:
          - type: number
          - type: 'null'
          title: Rating
        network:
          anyOf:
          - type: string
          - type: 'null'
          title: Network
        studio:
          anyOf:
          - type: string
          - type: 'null'
          title: Studio
      type: object
      required:
      - media_type
      - title
      - year
      - overview
      - already_in_library
      - external_id
      - poster_url
      - fanart_url
      - genres
      - runtime
      - rating
      - network
      - studio
      title: MediaSearchResultResponse
    MediaType:
      type: string
      enum:
      - series
      - movie
      title: MediaType
    RemoveMediaRequest:
      properties:
        media_type:
          $ref: '#/components/schemas/MediaType'
        id:
          type: integer
          title: Id
        delete_files:
          type: boolean
          title: Delete Files
          default: false
        add_exclusion:
          type: boolean
          title: Add Exclusion
          default: false
      type: object
      required:
      - media_type
      - id
      title: RemoveMediaRequest
    Token:
      properties:
        access_token:
          type: string
          title: Access Token
        token_type:
          type: string
          title: Token Type
      type: object
      required:
      - access_token
      - token_type
      title: Token
    UserCreate:
      properties:
        email:
          type: string
          format: email
          title: Email
        username:
          type: string
          title: Username
        password:
          type: string
          title: Password
      type: object
      required:
      - email
      - username
      - password
      title: UserCreate
    UserResponse:
      properties:
        id:
          type: integer
          title: Id
        email:
          type: string
          title: Email
        username:
          type: string
          title: Username
        is_active:
          type: boolean
          title: Is Active
        email_verified:
          type: boolean
          title: Email Verified
        role:
          type: string
          title: Role
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
      type: object
      required:
      - id
      - email
      - username
      - is_active
      - email_verified
      - role
      - created_at
      - updated_at
      title: UserResponse
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    OAuth2PasswordBearer:
      type: oauth2
      flows:
        password:
          scopes: {}
          tokenUrl: /api/auth/login